Collapse authorization SDK entrypoint file
