+++ sdk/go/authorization.go
@@ -1,9 +1,12 @@
 package gestalt
 
 import (
+	"context"
 	"time"
 
 	sdkauthorization "github.com/valon-technologies/gestalt/sdk/go/authorization"
+	rpcauthorization "github.com/valon-technologies/gestalt/server/rpc/authorization"
+	proto "github.com/valon-technologies/gestalt/server/rpc/protov1/v1"
 )
 
 type (
@@ -161,6 +164,25 @@ func NewAuthorizationModelTupleToUsersetRewrite(tuplesetRelation, computedRelati
 // NewAuthorizationModelUnionRewrite unions multiple rewrite branches.
 func NewAuthorizationModelUnionRewrite(children ...*AuthorizationModelRewrite) *AuthorizationModelRewrite {
 	return sdkauthorization.NewAuthorizationModelUnionRewrite(children...)
+}
+
+var sharedAuthorizationTransport sharedManagerTransport[proto.AuthorizationProviderClient]
+
+// Authorization returns a shared authorization capability.
+func Authorization() (sdkauthorization.Authorization, error) {
+	target, token, err := hostServiceTarget("authorization")
+	if err != nil {
+		return nil, err
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	client, err := managerTransportClient(ctx, "authorization", target, token, &sharedAuthorizationTransport, proto.NewAuthorizationProviderClient)
+	if err != nil {
+		return nil, err
+	}
+	return rpcauthorization.NewClient(client, rpcauthorization.Options{}), nil
 }
 
 // AuthorizationProvider serves authorization APIs to the host.
